all bugs on trips new page now removed
