+++ app/javascript/packs/application.js
@@ -7,7 +7,6 @@ require("@rails/ujs").start()
 require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-require('dotenv').config()
 
 
 
@@ -34,20 +33,18 @@ $('.popover-markup>.trigger').popover({
 
 // External imports
 import "bootstrap";
-import {displayinfo, getweatherdata } from '../getweather';
+console.log('hello');
 $(function () {
   $('[data-toggle="popover"]').popover({html: true})
 })
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
 import { initMapbox } from '../plugins/init_mapbox';
-
+import {displayinfo, getweatherdata } from '../plugins/getweather';
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
   // initSelect2();
-  if (document.getElementById('getweather')){
-    document.getElementById('getweather').addEventListener('click', () => {getweatherdata()})
-  }
+  getweatherdata();
   initMapbox();
 });
