+++ app/views/trips/edit.html.erb
@@ -1,0 +1,9 @@
+<h1>Edit this trip's logbook entry</h1>
+<p>Current entry:</p>
+<%= @trip.description %>
+<p>Updated entry:</p>
+<%= form_for(@trip) do |f| %>
+  <%= f.label :description %>
+  <%= f.text_field :description %>
+  <%= f.submit %>
+<% end %>
